requirements.txt
numpy
scikit-learn
face_recognition
pillow
tkinter